Refine bubble map: zoom-responsive sizes, hide no-data, default to Municipal
- Bubbles scale with zoom: 1–8px at state level (z6), 3–16px at metro (z10)
- Hide features with no data (radius/opacity 0) instead of showing gray dots
- Default agency type filter to Municipal
- Rename ChoroplethMap → AgencyMap
